fix: rename naming convention
